+++ src/physics/single_chain/ufjc/lennard_jones/thermodynamics/isotensional/legendre/test.rs
@@ -91,7 +91,7 @@ mod nondimensional
             let link_stiffness = parameters.link_stiffness_reference + parameters.link_stiffness_scale*(0.5 - rng.gen::<f64>());
             let model = LENNARDJONESFJC::init(number_of_links, link_length, hinge_mass, link_stiffness);
             let temperature = parameters.temperature_reference + parameters.temperature_scale*(0.5 - rng.gen::<f64>());
-            let lambda_max = (13.0/7.0 as f64).powf(1.0/6.0);
+            let lambda_max = (13.0/7.0_f64).powf(1.0/6.0);
             let nondimensional_force_max = link_stiffness/BOLTZMANN_CONSTANT/temperature*link_length.powi(2)/6.0*(lambda_max.powi(-7) - lambda_max.powi(-13));
             let nondimensional_force = nondimensional_force_max*rng.gen::<f64>();
             let nondimensional_helmholtz_free_energy = model.nondimensional_helmholtz_free_energy(&nondimensional_force, &temperature);
@@ -116,7 +116,7 @@ mod nondimensional
             let link_stiffness = parameters.link_stiffness_reference + parameters.link_stiffness_scale*(0.5 - rng.gen::<f64>());
             let model = LENNARDJONESFJC::init(number_of_links, link_length, hinge_mass, link_stiffness);
             let temperature = parameters.temperature_reference + parameters.temperature_scale*(0.5 - rng.gen::<f64>());
-            let lambda_max = (13.0/7.0 as f64).powf(1.0/6.0);
+            let lambda_max = (13.0/7.0_f64).powf(1.0/6.0);
             let nondimensional_force_max = link_stiffness/BOLTZMANN_CONSTANT/temperature*link_length.powi(2)/6.0*(lambda_max.powi(-7) - lambda_max.powi(-13));
             let nondimensional_force = nondimensional_force_max*rng.gen::<f64>();
             let nondimensional_helmholtz_free_energy_per_link = model.nondimensional_helmholtz_free_energy_per_link(&nondimensional_force, &temperature);
@@ -141,7 +141,7 @@ mod nondimensional
             let link_stiffness = parameters.link_stiffness_reference + parameters.link_stiffness_scale*(0.5 - rng.gen::<f64>());
             let model = LENNARDJONESFJC::init(number_of_links, link_length, hinge_mass, link_stiffness);
             let temperature = parameters.temperature_reference + parameters.temperature_scale*(0.5 - rng.gen::<f64>());
-            let lambda_max = (13.0/7.0 as f64).powf(1.0/6.0);
+            let lambda_max = (13.0/7.0_f64).powf(1.0/6.0);
             let nondimensional_force_max = link_stiffness/BOLTZMANN_CONSTANT/temperature*link_length.powi(2)/6.0*(lambda_max.powi(-7) - lambda_max.powi(-13));
             let nondimensional_force = nondimensional_force_max*rng.gen::<f64>();
             let nondimensional_relative_helmholtz_free_energy = model.nondimensional_relative_helmholtz_free_energy(&nondimensional_force, &temperature);
@@ -166,7 +166,7 @@ mod nondimensional
             let link_stiffness = parameters.link_stiffness_reference + parameters.link_stiffness_scale*(0.5 - rng.gen::<f64>());
             let model = LENNARDJONESFJC::init(number_of_links, link_length, hinge_mass, link_stiffness);
             let temperature = parameters.temperature_reference + parameters.temperature_scale*(0.5 - rng.gen::<f64>());
-            let lambda_max = (13.0/7.0 as f64).powf(1.0/6.0);
+            let lambda_max = (13.0/7.0_f64).powf(1.0/6.0);
             let nondimensional_force_max = link_stiffness/BOLTZMANN_CONSTANT/temperature*link_length.powi(2)/6.0*(lambda_max.powi(-7) - lambda_max.powi(-13));
             let nondimensional_force = nondimensional_force_max*rng.gen::<f64>();
             let nondimensional_relative_helmholtz_free_energy_per_link = model.nondimensional_relative_helmholtz_free_energy_per_link(&nondimensional_force, &temperature);
@@ -196,7 +196,7 @@ mod per_link
             let link_stiffness = parameters.link_stiffness_reference + parameters.link_stiffness_scale*(0.5 - rng.gen::<f64>());
             let model = LENNARDJONESFJC::init(number_of_links, link_length, hinge_mass, link_stiffness);
             let temperature = parameters.temperature_reference + parameters.temperature_scale*(0.5 - rng.gen::<f64>());
-            let lambda_max = (13.0/7.0 as f64).powf(1.0/6.0);
+            let lambda_max = (13.0/7.0_f64).powf(1.0/6.0);
             let nondimensional_force_max = link_stiffness/BOLTZMANN_CONSTANT/temperature*link_length.powi(2)/6.0*(lambda_max.powi(-7) - lambda_max.powi(-13));
             let nondimensional_force = nondimensional_force_max*rng.gen::<f64>();
             let force = nondimensional_force*BOLTZMANN_CONSTANT*temperature/link_length;
@@ -221,7 +221,7 @@ mod per_link
             let link_stiffness = parameters.link_stiffness_reference + parameters.link_stiffness_scale*(0.5 - rng.gen::<f64>());
             let model = LENNARDJONESFJC::init(number_of_links, link_length, hinge_mass, link_stiffness);
             let temperature = parameters.temperature_reference + parameters.temperature_scale*(0.5 - rng.gen::<f64>());
-            let lambda_max = (13.0/7.0 as f64).powf(1.0/6.0);
+            let lambda_max = (13.0/7.0_f64).powf(1.0/6.0);
             let nondimensional_force_max = link_stiffness/BOLTZMANN_CONSTANT/temperature*link_length.powi(2)/6.0*(lambda_max.powi(-7) - lambda_max.powi(-13));
             let nondimensional_force = nondimensional_force_max*rng.gen::<f64>();
             let force = nondimensional_force*BOLTZMANN_CONSTANT*temperature/link_length;
@@ -246,7 +246,7 @@ mod per_link
             let link_stiffness = parameters.link_stiffness_reference + parameters.link_stiffness_scale*(0.5 - rng.gen::<f64>());
             let model = LENNARDJONESFJC::init(number_of_links, link_length, hinge_mass, link_stiffness);
             let temperature = parameters.temperature_reference + parameters.temperature_scale*(0.5 - rng.gen::<f64>());
-            let lambda_max = (13.0/7.0 as f64).powf(1.0/6.0);
+            let lambda_max = (13.0/7.0_f64).powf(1.0/6.0);
             let nondimensional_force_max = link_stiffness/BOLTZMANN_CONSTANT/temperature*link_length.powi(2)/6.0*(lambda_max.powi(-7) - lambda_max.powi(-13));
             let nondimensional_force = nondimensional_force_max*rng.gen::<f64>();
             let nondimensional_helmholtz_free_energy = model.nondimensional_helmholtz_free_energy(&nondimensional_force, &temperature);
@@ -270,7 +270,7 @@ mod per_link
             let link_stiffness = parameters.link_stiffness_reference + parameters.link_stiffness_scale*(0.5 - rng.gen::<f64>());
             let model = LENNARDJONESFJC::init(number_of_links, link_length, hinge_mass, link_stiffness);
             let temperature = parameters.temperature_reference + parameters.temperature_scale*(0.5 - rng.gen::<f64>());
-            let lambda_max = (13.0/7.0 as f64).powf(1.0/6.0);
+            let lambda_max = (13.0/7.0_f64).powf(1.0/6.0);
             let nondimensional_force_max = link_stiffness/BOLTZMANN_CONSTANT/temperature*link_length.powi(2)/6.0*(lambda_max.powi(-7) - lambda_max.powi(-13));
             let nondimensional_force = nondimensional_force_max*rng.gen::<f64>();
             let nondimensional_relative_helmholtz_free_energy = model.nondimensional_relative_helmholtz_free_energy(&nondimensional_force, &temperature);
@@ -299,7 +299,7 @@ mod relative
             let link_stiffness = parameters.link_stiffness_reference + parameters.link_stiffness_scale*(0.5 - rng.gen::<f64>());
             let model = LENNARDJONESFJC::init(number_of_links, link_length, hinge_mass, link_stiffness);
             let temperature = parameters.temperature_reference + parameters.temperature_scale*(0.5 - rng.gen::<f64>());
-            let lambda_max = (13.0/7.0 as f64).powf(1.0/6.0);
+            let lambda_max = (13.0/7.0_f64).powf(1.0/6.0);
             let nondimensional_force_max = link_stiffness/BOLTZMANN_CONSTANT/temperature*link_length.powi(2)/6.0*(lambda_max.powi(-7) - lambda_max.powi(-13));
             let nondimensional_force = nondimensional_force_max*rng.gen::<f64>();
             let force = nondimensional_force*BOLTZMANN_CONSTANT*temperature/link_length;
@@ -325,7 +325,7 @@ mod relative
             let link_stiffness = parameters.link_stiffness_reference + parameters.link_stiffness_scale*(0.5 - rng.gen::<f64>());
             let model = LENNARDJONESFJC::init(number_of_links, link_length, hinge_mass, link_stiffness);
             let temperature = parameters.temperature_reference + parameters.temperature_scale*(0.5 - rng.gen::<f64>());
-            let lambda_max = (13.0/7.0 as f64).powf(1.0/6.0);
+            let lambda_max = (13.0/7.0_f64).powf(1.0/6.0);
             let nondimensional_force_max = link_stiffness/BOLTZMANN_CONSTANT/temperature*link_length.powi(2)/6.0*(lambda_max.powi(-7) - lambda_max.powi(-13));
             let nondimensional_force = nondimensional_force_max*rng.gen::<f64>();
             let force = nondimensional_force*BOLTZMANN_CONSTANT*temperature/link_length;
@@ -351,7 +351,7 @@ mod relative
             let link_stiffness = parameters.link_stiffness_reference + parameters.link_stiffness_scale*(0.5 - rng.gen::<f64>());
             let model = LENNARDJONESFJC::init(number_of_links, link_length, hinge_mass, link_stiffness);
             let temperature = parameters.temperature_reference + parameters.temperature_scale*(0.5 - rng.gen::<f64>());
-            let lambda_max = (13.0/7.0 as f64).powf(1.0/6.0);
+            let lambda_max = (13.0/7.0_f64).powf(1.0/6.0);
             let nondimensional_force_max = link_stiffness/BOLTZMANN_CONSTANT/temperature*link_length.powi(2)/6.0*(lambda_max.powi(-7) - lambda_max.powi(-13));
             let nondimensional_force = nondimensional_force_max*rng.gen::<f64>();
             let nondimensional_helmholtz_free_energy = model.nondimensional_helmholtz_free_energy(&nondimensional_force, &temperature);
@@ -376,7 +376,7 @@ mod relative
             let link_stiffness = parameters.link_stiffness_reference + parameters.link_stiffness_scale*(0.5 - rng.gen::<f64>());
             let model = LENNARDJONESFJC::init(number_of_links, link_length, hinge_mass, link_stiffness);
             let temperature = parameters.temperature_reference + parameters.temperature_scale*(0.5 - rng.gen::<f64>());
-            let lambda_max = (13.0/7.0 as f64).powf(1.0/6.0);
+            let lambda_max = (13.0/7.0_f64).powf(1.0/6.0);
             let nondimensional_force_max = link_stiffness/BOLTZMANN_CONSTANT/temperature*link_length.powi(2)/6.0*(lambda_max.powi(-7) - lambda_max.powi(-13));
             let nondimensional_force = nondimensional_force_max*rng.gen::<f64>();
             let nondimensional_helmholtz_free_energy_per_link = model.nondimensional_helmholtz_free_energy_per_link(&nondimensional_force, &temperature);
